Make sure decoding crashes doesn't... crash
